added 4 Sum Problem
We are using the Naïve Recursive Approach to solve this problem.The idea is similar to the 0–1 Knapsack problem and uses recursion. For each item, either consider it or exclude it and recur for the remaining items. Return true if the desired sum is found by including or excluding the current item.